requirements.txt
langchain>=0.1.0
langchain-google-genai>=1.0.0
langchain-mistralai>=0.1.0
langchain-openai>=0.1.0
python-dotenv>=1.0.0
streamlit>=1.28.0
gradio>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
jupyter>=1.0.0
psutil>=5.9.0
tiktoken>=0.5.0
pydantic>=2.0.0
python-json-logger>=2.0.0
